project_smartfly/pyproject.toml
[project]
name = "project-smartfly"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.11.0.86",
    "semantic-kernel>=1.29.0",
    "colorlog>=6.8.0",
    "python-json-logger>=2.0.7",
    "dependencies>=7.7.1",
    "flask>=3.1.0",
    "flask-socketio>=5.5.1",
    "pyserial>=3.5",
    "google>=3.0.0",
    "google-generativeai>=0.8.5",
    "crewai>=0.119.0",
    "crewai-tools>=0.44.0",
    "groq>=0.24.0",
    "google-genai>=1.14.0",
]
